pyproject.toml
[project]
name = "dataflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-beam[gcp]>=2.67.0",
    "faker>=37.6.0",
    "pandas>=2.3.2",
    "pip>=25.2",
    "python-dotenv>=1.1.1",
]
